backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
openai>=1.0.0
anthropic>=0.7.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
pydantic>=2.0.0